requirements.txt
altair==5.4.1
aniso8601==8.0.0
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
Authlib==1.3.2
babel==2.16.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
blis==1.0.1
branca==0.8.0
bs4==0.0.2
cachelib==0.13.0
cachetools==5.5.0
captcha==0.6.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==2.0.1
extra-streamlit-components==0.1.71
fastjsonschema==2.20.0
Flask==3.0.3
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
Flask-RESTful==0.3.8
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
folium==0.18.0
fonttools==4.54.1
fqdn==1.5.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
googlemaps==4.10.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
msgspec==0.18.6
multipledispatch==1.0.0
murmurhash==1.0.10
mysql-connector-python==9.0.0
mysqlclient==2.2.4
narwhals==1.13.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.0
oauthlib==3.2.2
openai==1.54.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
preshed==3.0.9
prettytable==3.11.0
prometheus_client==0.21.0
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.0.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2020.1
pywin32==306
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.25.0
Send2Trash==1.8.3
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
stack-data==0.6.3
streamlit==1.39.0
streamlit-authenticator==0.4.1
streamlit-router==0.1.8
streamlit_folium==0.23.1
tenacity==9.0.0
terminado==0.18.1
thinc==8.3.2
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typer==0.13.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
URLObject==2.4.3
wasabi==1.1.3
watchdog==5.0.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.45.0
wrapt==1.16.0
wsproto==1.2.0
xyzservices==2024.9.0
